feat(ng-draw-flow): support Angular 19 and ng-web-apis 5

+++ projects/ng-draw-flow/src/lib/components/connectors/base-connector.ts
@@ -1,6 +1,4 @@
-import {DestroyRef, Directive, ElementRef, HostBinding, inject} from '@angular/core';
-import {takeUntilDestroyed} from '@angular/core/rxjs-interop';
-import {filter} from 'rxjs';
+import {Directive, effect, ElementRef, inject} from '@angular/core';
 
 import {
     type DfConnectionPoint,
@@ -10,50 +8,58 @@ import {
 } from '../../ng-draw-flow.interfaces';
 import {ConnectionsService} from '../connections/connections.service';
 
-@Directive()
+@Directive({
+    host: {
+        '[attr.data-connector-type]': 'connectorType',
+        '[attr.data-node-id]': 'bindNodeId',
+        '[attr.data-connector-id]': 'bindConnectorId',
+        '[attr.data-position]': 'bindPosition',
+    },
+})
 export abstract class BaseConnector {
-    protected abstract data: DfDataConnectorConfig;
-
-    // eslint-disable-next-line @angular-eslint/prefer-host-metadata-property
-    @HostBinding('attr.data-connector-type')
     protected connectorType!: DfConnectionPoint;
 
-    protected readonly destroyRef = inject(DestroyRef);
     protected isDisabled = false;
     protected readonly connectionsService = inject(ConnectionsService);
 
-    protected readonly sub = this.connectionsService.usedConnectors$
-        .pipe(
-            filter(() => !!this.data.connectorId),
-            takeUntilDestroyed(),
-        )
-        .subscribe((usedConnectorIds: string[]) => {
-            this.setupDisabledState(usedConnectorIds.includes(this.data.connectorId));
-        });
-
     public coordinates?: DfPoint;
-    public position?: DfConnectorPosition;
     public readonly nativeElement = inject(ElementRef).nativeElement;
 
-    // eslint-disable-next-line @angular-eslint/prefer-host-metadata-property
-    @HostBinding('attr.data-node-id')
-    public get bindNodeId(): string {
-        return this.data.nodeId;
+    constructor() {
+        effect(() => {
+            const connectorId = this.data?.connectorId;
+
+            if (!connectorId) {
+                return;
+            }
+
+            this.setupDisabledState(
+                this.connectionsService.usedConnectors().includes(connectorId),
+            );
+        });
     }
 
-    // eslint-disable-next-line @angular-eslint/prefer-host-metadata-property
-    @HostBinding('attr.data-connector-id')
-    public get bindConnectorId(): string {
-        return this.data.connectorId;
+    public abstract get position(): DfConnectorPosition | undefined;
+
+    protected abstract get data(): DfDataConnectorConfig;
+
+    public get bindNodeId(): string | undefined {
+        return this.data?.nodeId;
     }
 
-    // eslint-disable-next-line @angular-eslint/prefer-host-metadata-property
-    @HostBinding('attr.data-position')
+    public get bindConnectorId(): string | undefined {
+        return this.data?.connectorId;
+    }
+
     public get bindPosition(): DfConnectorPosition | undefined {
         return this.position;
     }
 
     public destroy(): void {
+        if (!this.data?.connectorId) {
+            return;
+        }
+
         this.connectionsService.removeConnectionsByConnectorId(this.data.connectorId);
     }
 
@@ -64,7 +70,7 @@ export abstract class BaseConnector {
             this.nativeElement.removeAttribute('data-connected');
         }
 
-        this.isDisabled = this.data.single && connected;
+        this.isDisabled = (this.data?.single satisfies boolean) && connected;
 
         this.nativeElement.classList.toggle('df-disabled', this.isDisabled);
     }
